Upgraded to Microchip's latest project overhaul
Ahead of releasing the AVR-IoG WG board, Microchip replaced it's
demo application on Atmel START with an overhauled revision. The
functionality is the same, but the project structure and middleware
files have been changed along with the cloud_TASK implementation. There
are likely many other changes I have yet to notice. I just replaced
every file in my project with the files in the updated project,
effectively starting over.