Re-configured Atmel START project
I re-configured the initial level of the RFID_CLICK_SPI_CS and
RFID_CLICK_SSI0 to be high. This just makes it easier as the function
to initialize the click board doesn't get called until the rest of the
system is initialized (it can't come before because the SPI bus needs to
be initialized and I have not added that to the click board init
function).