Modified demo to read and send tag UID to GC
Rather than reading sensor data every second and sending it to Google
Cloud, the program now scans for an RFID tag every second and, if one is
present and it's UID is successfully read, its UID is sent to Google
Cloud. I've decided to leave the yellow LED alone and let it blink
everytime an RFID scan takes place.

Note that I would rather have performed a scan every half second, but as
the main_datatask uses time.h to check the time delta, there is only a
resolution of one second. I may modify this in the future, but it is not
worth my time right now.

I have successfully tested these changes using a Google Cloud function
that checks the UID string against a database. If the UID is in the
database, it sends the device a "YES" command. Otherwise, the device
recieves a "NO" command.