Ingetrated Microchip's latest project update
After I reconfigured the SMART project to initialize the
ACCESS_CONTROL_PIN, the project failed to build because some of the
files were updated with the lastest libraries and others were not (the
ones I modified). So I merged those updates manually and re-tested the
application to confirm its functionality.